Circles for winners

+++ rivalryapp/AppDelegate.m
@@ -41,7 +41,7 @@
     {
         if ([notificaiton.allKeys indexOfObject:@"contentUrl"] != NSNotFound)
         {
-            dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(1.0 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
+            dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (1.0 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
                 [[UIApplication sharedApplication] openURL:[NSURL URLWithString:notificaiton[@"contentUrl"]]];
                 exit(0);
             });
